JSON and TSV data, with scripts.
This provides Python code that converts the data contained in README.md into JSON and TSV formats.

Start with scripts/init.sh to initialize the Python virtual environment.

Then use scripts/build.sh to extract the data.

Copies of the data are included in this commit. We should probably add a commit hook to build automatically.