The September 1942 Improvement Era.
By far the largest Improvement Era batch so far. It is clear that the church's information processing and communications capabilities matched the technology of the era, given the delay between the actions and the reporting of the actions, and apparent inaccuracies in the information which a quick Google search can reveal.

Entries now total 245.

This introduces `scripts/resort.py` which intelligently reorders the elements in discipline_base.json.